MAJOR: Change QC_min_q30 to 80 instead of 90 and have just_like_2024 affect it
myco_sra now has the same default QC_min_q30 as myco_raw. Additionally, since the main Q30 filter was effectively a no-op in the 2024 version, that behavior is now replicated in myco_raw AND myco_sra when just_like_2024. This commit doesn't touch preliminary_min_q30 although it may be worth investigating if that should also be changed when just_like_2024 and !guardrail_mode.